int 数据点都用unit64 接

+++ main/boards/gizwits-c2-6824-lvlin/lvlin_data_point_manager.cc
@@ -223,7 +223,7 @@ size_t LvlinDataPointManager::GetDataPointCount() const {
 }
 
 // 标准实现：获取数据点值
-bool LvlinDataPointManager::GetDataPointValue(const std::string& name, int& value) const {
+bool LvlinDataPointManager::GetDataPointValue(const std::string& name, uint32_t& value) const {
     if (name == "switch") {
         value = 1; // 开关状态，固定为1
         return true;
@@ -312,26 +312,26 @@ bool LvlinDataPointManager::GetDataPointValue(const std::string& name, int& valu
 }
 
 // 标准实现：设置数据点值
-bool LvlinDataPointManager::SetDataPointValue(const std::string& name, int value) {
+bool LvlinDataPointManager::SetDataPointValue(const std::string& name, uint32_t value) {
     if (name == "chat_mode") {
         if (set_chat_mode_callback_) {
-            set_chat_mode_callback_(value);
+            set_chat_mode_callback_(static_cast<int>(value));
             return true;
         }
     } else if (name == "volume_set") {
         if (set_volume_callback_) {
-            set_volume_callback_(value);
+            set_volume_callback_(static_cast<int>(value));
             return true;
         }
     } else if (name == "brightness") {
         if (set_brightness_callback_) {
-            set_brightness_callback_(value);
+            set_brightness_callback_(static_cast<int>(value));
             return true;
         }
     } else if (name == "speed") {
         if (set_speed_callback_) {
             // 限制语速值在有效范围内 (0-200, 对应-50%到150%)
-            int clamped_value = std::max(0, std::min(200, value));
+            int clamped_value = std::max(0, std::min(200, static_cast<int>(value)));
             set_speed_callback_(clamped_value);
             return true;
         }
@@ -422,8 +422,8 @@ void LvlinDataPointManager::GenerateReportData(uint8_t* buffer, size_t buffer_si
 }
 
 // 标准实现：处理数据点值
-void LvlinDataPointManager::ProcessDataPointValue(const std::string& name, int value) {
-    ESP_LOGI(TAG, "ProcessDataPointValue: %s = %d", name.c_str(), value);
+void LvlinDataPointManager::ProcessDataPointValue(const std::string& name, uint32_t value) {
+    ESP_LOGI(TAG, "ProcessDataPointValue: %s = %u", name.c_str(), value);
     SetDataPointValue(name, value);
 }
